Remove getDeadline deprecation warning and refactor the devtools package
