+++ packages/devtools/package.json
