+++ packages/devtools/src/devtools/ui/DevtoolsLayout.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { DevtoolsAgentProvider } from "./Context";
+import { DevtoolsClientProvider } from "./DevtoolsClientProvider";
 
 type LayoutProps = {
   sidebar: React.ReactNode;
@@ -7,11 +7,11 @@ type LayoutProps = {
 };
 export function DevtoolsLayout(props: LayoutProps) {
   return (
-    <DevtoolsAgentProvider>
+    <DevtoolsClientProvider>
       <div className="asd-root">
         <div className="asd-side asc-s-y-auto">{props.sidebar}</div>
         <div className="asd-view asc-s-y-auto">{props.details}</div>
       </div>
-    </DevtoolsAgentProvider>
+    </DevtoolsClientProvider>
   );
 }
