+++ packages/devtools/src/static/background.js
