+++ packages/devtools/src/static/contentScript.js
