+++ packages/devtools/src/static/main.js
