+++ packages/devtools/src/static/manifest.json
