+++ packages/devtools/vite.config.lib.prod.ts
